requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
future==0.18.3
idna==3.7
py-linq==1.4.0
python-dateutil==2.9.0.post0
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
smartsheet-python-sdk==3.0.2
urllib3==2.2.1
